feat: add footer and refactor social links
